Particle Unit Conversion: Generalize
Generalize the member to convert the internal proper velocity
to SI momentum and back.

This is needed for BTD diagnostics, where we have a less specialized
but pinned AMReX particle container, which does not have this method.
