+++ Source/Particles/PhysicalParticleContainer.H
@@ -234,6 +234,12 @@ public:
         const amrex::Real t_lab, const amrex::Real dt,
         DiagnosticParticles& diagnostic_particles) final;
 
+    /** Convert particle momentum to/from SI
+     *
+     * @see particlesConvertUnits
+     *
+     * @param convert_dir convert to or from SI
+     */
     virtual void ConvertUnits (ConvertDirection convert_dir) override;
 
 /**
