+++ Source/Particles/WarpXParticleContainer.H
@@ -42,38 +42,6 @@
 #include <string>
 #include <utility>
 
-enum struct ConvertDirection{WarpX_to_SI, SI_to_WarpX};
-
-struct PIdx
-{
-    enum { // Particle Attributes stored in amrex::ParticleContainer's struct of array
-        w = 0,  // weight
-        ux, uy, uz,
-#ifdef WARPX_DIM_RZ
-        theta, // RZ needs all three position components
-#endif
-        nattribs
-    };
-};
-
-struct DiagIdx
-{
-    enum {
-        w = 0,
-        x, y, z, ux, uy, uz,
-        nattribs
-    };
-};
-
-struct TmpIdx
-{
-    enum {
-        xold = 0,
-        yold, zold, uxold, uyold, uzold,
-        nattribs
-    };
-};
-
 namespace ParticleStringNames
 {
     const std::map<std::string, int> to_index = {
